opsmedic_project/buggy_app_v2/requirements.txt
flask
gunicorn
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-distro